planet_lib_requirements.txt
affine==2.3.1
anyio==3.6.1
args==0.1.0
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
clint==0.5.1
colorama==0.4.5
coverage==6.2
docopt==0.6.2
Fiona @ file:///C:/Users/P/pipwin/Fiona-1.8.21-cp310-cp310-win_amd64.whl
GDAL @ file:///C:/Users/P/pipwin/GDAL-3.4.3-cp310-cp310-win_amd64.whl
geojson==2.5.0
geopandas==0.11.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
Js2Py==0.71
jsonschema==4.14.0
mamba==0.11.2
munch==2.5.0
numpy @ file:///C:/Users/P/pipwin/numpy-1.22.4%2Bvanilla-cp310-cp310-win_amd64.whl
packaging==21.3
pandas==1.4.3
pipwin==0.5.2
planet==2.0a2
pyjsparser==2.7.1
PyJWT==2.4.0
pyparsing==3.0.9
PyPrind==2.11.3
pyproj @ file:///C:/Users/P/pipwin/pyproj-3.3.1-cp310-cp310-win_amd64.whl
pyrsistent==0.18.1
pySmartDL==1.3.4
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
rasterio==1.3.2
requests==2.28.1
rfc3986==1.5.0
Rtree @ file:///C:/Users/P/pipwin/Rtree-1.0.0-cp310-cp310-win_amd64.whl
Shapely @ file:///C:/Users/P/pipwin/Shapely-1.8.2-cp310-cp310-win_amd64.whl
six==1.16.0
sniffio==1.2.0
snuggs==1.4.7
soupsieve==2.3.2.post1
tqdm==4.64.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
